refactor: avatarItem 패키지 이름을 avatar로 변경
