+++ src/main/java/com/backendoori/ootw/avtar/domain/Type.java
@@ -1,4 +1,4 @@
-package com.backendoori.ootw.avtarItem.domain;
+package com.backendoori.ootw.avtar.domain;
 
 public enum Type {
     HAIR, TOP, PANTS, ACCESSORY, SHOES, BACKGROUND
