+++ src/main/java/com/backendoori/ootw/avtar/repository/AvatarItemRepository.java
@@ -1,6 +1,6 @@
-package com.backendoori.ootw.avtarItem.repository;
+package com.backendoori.ootw.avtar.repository;
 
-import com.backendoori.ootw.avtarItem.domain.AvatarItem;
+import com.backendoori.ootw.avtar.domain.AvatarItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AvatarItemRepository extends JpaRepository<AvatarItem, Long> {
